requirements.txt
ultralytics>=8.1.0
opencv-python-headless>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
yt-dlp>=2024.1.0
requests>=2.31.0
tqdm>=4.66.0
